Refactoring

+++ kernel/src/processor.rs
@@ -16,7 +16,7 @@ pub struct Controller {
 impl Controller {
     pub fn boot(&self, boot_loader: &mut boot::Loader, local_apic_registers: &mut interrupt::apic::local::Registers, hpet: &timer::hpet::Registers) {
         boot_loader.initialize();
-        let local_apic_id: u8 = self.local_apic_id() as u8;
+        let local_apic_id: u8 = self.local_apic_id();
         com2_println!("Boot processor {:#x?}", local_apic_id);
         let entry_point: usize = boot_loader.entry_point();
         local_apic_registers.send_init(local_apic_id, hpet);
